+++ public/Nitrocid.Addons/Nitrocid.Extras.ColorConvert/Commands/ColorYuvToRgb.cs
@@ -38,17 +38,17 @@ namespace Nitrocid.Extras.ColorConvert.Commands
         public override int Execute(CommandParameters parameters, ref string variableValue)
         {
             // Check to see if we have the numeric arguments
-            if (!double.TryParse(parameters.ArgumentsList[0], out double Y))
+            if (!int.TryParse(parameters.ArgumentsList[0], out int Y))
             {
                 TextWriters.Write(Translate.DoTranslation("The Y component level must be numeric."), true, KernelColorType.Error);
                 return 10000 + (int)KernelExceptionType.Color;
             }
-            if (!double.TryParse(parameters.ArgumentsList[1], out double U))
+            if (!int.TryParse(parameters.ArgumentsList[1], out int U))
             {
                 TextWriters.Write(Translate.DoTranslation("The U component level must be numeric."), true, KernelColorType.Error);
                 return 10000 + (int)KernelExceptionType.Color;
             }
-            if (!double.TryParse(parameters.ArgumentsList[2], out double V))
+            if (!int.TryParse(parameters.ArgumentsList[2], out int V))
             {
                 TextWriters.Write(Translate.DoTranslation("The V component level must be numeric."), true, KernelColorType.Error);
                 return 10000 + (int)KernelExceptionType.Color;
